:recycle: refactor : style코드 분리
